add log levels map
